Add files via upload
This project is called Ojay-Analysis-tool its a generic desktop analysis tool which can be used to predict, analyze and generate a graphical visual report to check the prevalence of malaria, typhoid and fever. This tool will also verify, if two variables (gender and diagnosis) are related to each other. Pearson Chi-square formular was implemented for testing variables. This application is purely written in python.
I, had a tough time at some point developing this application but I guess it was what the wait.
Note take note of the steps and images uploaded.